split tests into different packages
