+++ tests/server/watch/leader_watch_test.go
@@ -11,10 +11,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package server_test
+package watch_test
 
 import (
 	"context"
+	"testing"
 	"time"
 
 	. "github.com/pingcap/check"
@@ -23,6 +24,18 @@ import (
 	"github.com/pingcap/pd/server"
 	"github.com/pingcap/pd/tests"
 )
+
+func Test(t *testing.T) {
+	TestingT(t)
+}
+
+var _ = Suite(&serverTestSuite{})
+
+type serverTestSuite struct{}
+
+func (s *serverTestSuite) SetUpSuite(c *C) {
+	server.EnableZap = true
+}
 
 func (s *serverTestSuite) TestWatcher(c *C) {
 	c.Parallel()
